Register & move worker functions into submodules
Use singledispatch to invoke them.
