+++ pybroom/__init__.py
@@ -1,0 +1,6 @@
+from .pybroom import tidy, glance, augment
+
+from .lmfit import lmfit  # noqa F401
+from .scipy import scipy  # noqa F401
+
+__all__ = ['tidy', 'glance', 'augment']
